Refactor Toast component to use NotificationService and update styles
