+++ examples/Demo/FluentUI.Demo.Client/Documentation/Components/Toast/Examples/FluentToastCustomDismiss.razor
@@ -1,4 +1,4 @@
-﻿@inject IToastService ToastService
+﻿@* @inject IToastService ToastService
 
 <FluentButton OnClick="@OpenToastAsync">
     Make toast
@@ -30,4 +30,4 @@
         Console.WriteLine($"Toast result: {result}");
     }
 }
-
+ *@
