+++ examples/Demo/FluentUI.Demo.Client/Documentation/Components/Toast/Examples/FluentToastInverted.razor
@@ -1,4 +1,4 @@
-﻿@inject IToastService ToastService
+﻿@* @inject IToastService ToastService
 
 <FluentButton OnClick="@OpenToastAsync">
    Make toast
@@ -32,4 +32,4 @@
         Console.WriteLine($"Toast result: {result}");
     }
 }
-
+ *@
